core/database/src/main/java/com/github/guibrisson/tucano/database/dao/TaskDao.kt
package com.github.guibrisson.tucano.database.dao

import androidx.room.Dao
import androidx.room.Delete
import androidx.room.Insert
import androidx.room.Query
import androidx.room.Upsert
import com.github.guibrisson.tucano.database.model.TaskEntity
import kotlinx.coroutines.flow.Flow

@Dao
interface TaskDao {
    @Query("SELECT * FROM tasks")
    fun getAllTasks(): Flow<List<TaskEntity>>

    @Query("SELECT * FROM tasks WHERE id = :id")
    fun getTaskById(id: String): Flow<TaskEntity?>

    @Upsert
    suspend fun upsertTask(vararg task: TaskEntity)

    @Delete
    suspend fun delete(vararg task: TaskEntity)
}
